Better kotlin tests

+++ project/jimmer-sql-kotlin/src/test/kotlin/org/babyfish/jimmer/sql/kt/filter/DeleteTest.kt
@@ -1,5 +1,6 @@
 package org.babyfish.jimmer.sql.kt.filter
 
+import org.babyfish.jimmer.sql.ast.impl.mutation.save.QueryReason
 import org.babyfish.jimmer.sql.kt.common.AbstractMutationTest
 import org.babyfish.jimmer.sql.kt.filter.common.FileFilter
 import org.babyfish.jimmer.sql.kt.model.filter.File
@@ -18,8 +19,14 @@ class DeleteTest : AbstractMutationTest() {
                 _sqlClient.entities.delete(File::class, 8L, it)
             }) {
                 statement {
-                    sql("delete from FILE_USER_MAPPING where FILE_ID = ?")
+                    sql("select USER_ID from FILE_USER_MAPPING where FILE_ID = ?")
                     variables(8L)
+                    queryReason(QueryReason.UPSERT_NOT_SUPPORTED)
+                }
+                statement {
+                    sql("delete from FILE_USER_MAPPING where FILE_ID = ? and USER_ID = ?")
+                    batchVariables(0, 8L, 2L)
+                    batchVariables(1, 8L, 3L)
                 }
                 statement {
                     sql(
@@ -30,6 +37,7 @@ class DeleteTest : AbstractMutationTest() {
                             |where tb_3_.PARENT_ID = ?""".trimMargin()
                     )
                     variables(8L)
+                    queryReason(QueryReason.TOO_DEEP)
                 }
                 statement {
                     sql(
@@ -57,14 +65,27 @@ class DeleteTest : AbstractMutationTest() {
                 }
                 statement {
                     sql(
-                        """delete from FILE_USER_MAPPING tb_1_ 
-                            |where exists (
-                            |--->select * 
-                            |--->from FILE tb_2_ 
-                            |--->where tb_1_.FILE_ID = tb_2_.ID and tb_2_.PARENT_ID = ?
-                            |)""".trimMargin()
+                        "select FILE_ID, USER_ID " +
+                            "from FILE_USER_MAPPING tb_1_ " +
+                            "inner join FILE tb_2_ on tb_1_.FILE_ID = tb_2_.ID " +
+                            "where tb_2_.PARENT_ID = ?"
                     )
                     variables(8L)
+                }
+                statement {
+                    sql(
+                        """delete from FILE_USER_MAPPING where FILE_ID = ? and USER_ID = ?""".trimMargin()
+                    )
+                    batchVariables(0, 9L, 2L)
+                    batchVariables(1, 9L, 3L)
+                    batchVariables(2, 10L, 3L)
+                    batchVariables(3, 10L, 4L)
+                    batchVariables(4, 11L, 2L)
+                    batchVariables(5, 11L, 4L)
+                    batchVariables(6, 12L, 2L)
+                    batchVariables(7, 12L, 3L)
+                    batchVariables(8, 13L, 3L)
+                    batchVariables(9, 13L, 4L)
                 }
                 statement {
                     sql("delete from FILE where PARENT_ID = ?")
